TIP-860: Move PoductModelNormalizer for Versioning from Tool to Pim/Enrichment
