+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/serializers_versionning.yml
@@ -2,6 +2,7 @@ parameters:
     pim_serializer.normalizer.flat.category.class: Akeneo\Pim\Enrichment\Component\Category\Normalizer\Versioning\CategoryNormalizer
     pim_serializer.normalizer.flat.product.class: Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\ProductNormalizer
     pim_serializer.normalizer.flat.value.class: Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product\ValueNormalizer
+    pim_serializer.normalizer.flat.product_model.class:     Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\ProductModelNormalizer
 
 services:
     pim_versioning.serializer.normalizer.flat.category:
@@ -20,5 +21,10 @@ services:
     pim_versioning.serializer.normalizer.flat.value:
         public: false
         class: '%pim_serializer.normalizer.flat.value.class%'
+        tags:
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    pim_versioning.serializer.normalizer.flat.product_model:
+        class: '%pim_serializer.normalizer.flat.product_model.class%'
         tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
