+++ src/Akeneo/Tool/Bundle/VersioningBundle/Resources/config/serializer/flat.yml
@@ -3,7 +3,6 @@ parameters:
     pim_serializer.normalizer.flat.group.class:             Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\GroupNormalizer
     pim_serializer.normalizer.flat.job_instance.class:      Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\JobInstanceNormalizer
     pim_serializer.normalizer.flat.label_translation.class: Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\TranslationNormalizer
-    pim_serializer.normalizer.flat.product_model.class:     Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\ProductModelNormalizer
     pim_serializer.normalizer.flat.collection.class:        Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\CollectionNormalizer
     pim_serializer.normalizer.flat.datetime.class:          Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\DateTimeNormalizer
     pim_serializer.normalizer.flat.metric.class:            Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\MetricNormalizer
@@ -11,11 +10,6 @@ parameters:
     pim_serializer.normalizer.flat.file.class:              Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\FileNormalizer
 
 services:
-    pim_versioning.serializer.normalizer.flat.product_model:
-        class: '%pim_serializer.normalizer.flat.product_model.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
     pim_versioning.serializer.normalizer.flat.group:
         class: '%pim_serializer.normalizer.flat.group.class%'
         arguments:
